Rename M0 to 'FORTH and add TARGET-'FORTH
Avoid confusion with Beetle’s M0, which has a different meaning, and
give a name to the constant representing the base address of the host
system.

The process uncovered a bug in Beetle’s meta.fs: IMMEDIATE should not be
applied to redefined words, as they are headerless, so IMMEDIATE will
apply to the previously defined word. This didn’t matter before, as that
word was never compiled.