Reorganise sources to isolate platform compiler routines
This is a prelude to using the platform compiler routines directly in
the meta-compiler, rather than duplicating them.

To this end, split machdeps.fs into compiler.fs and terminal.fs (the
latter is INCLUDEd by highlevel.fs just after the former), and rename
machdeps-pre.fs to os-primitives.fs.

The reordering has changed the RISC OS build, but not the Beetle build,
which has fewer OS access words.