Replace metacompiler with full compiler
Instead of redefining a few key words to rewire the current compiler,
load a full(er) compiler, splitting out the definitions from
highlevel.fs. This avoids the duplication of the metacompiler approach,
and is less fragile in the face of compiler changes.

As a result, fold the two meta.fs files into make.fs, splitting out a
few system-specific bits into other files.

Make FIND-AND-COMPILE, take a counted string, rather than a pointer to
an xt, so that it can be used more widely, and actually does what its
name suggests.