Autoconfiscate
Use a wrapper script, we can’t assume binfmt_misc support, cool as that is!

Build for both platforms, and then rebuild with the new compiler to check it
produces identical output. (This will be desired more often than not; it can
easily be manually overridden when e.g. the binary format or code generation
changes.)