Split system into a metacompiled base target, and an extended target
Add the assembler to the extended target; at the moment we can build this
only for Beetle.

Artificially serialize the builds of the various flavours of image, as
otherwise the build system may attempt (for a VPATH build) to build them
simultaneously in the same directory, resulting in file clashes.