Use build and host system setting
This prepares the way for multiple VM back-ends, and potentially (very
weakly potentially!) makes the build system RISC OS-compatible (if
everything else about it works, which is unlikely).

Allow the host and build system to be detected or set, and do not attempt to
triple-test or build the extended pforth.img unless they are the same.

Build all targets in any case, so we can keep git up to date, and keep
checking that changes that should not change the output do not on any back
end.