RISC OS: reorder primitives to fix bad VARIABLE
'THROW was defined at the top of the file, before (CREATE). This didn’t
cause a problem because the variable is not directly called, but of course
it meant that calling it caused a crash.

I noticed the problem when building pForth with SMite caused the incorrect
call of the host (CREATE) to change address and hence change pforth-base.img
when it should not have changed.