Remove some assumptions in the meta-compiler
Move WILL-DO into its own file, so it can be included as part of the
meta-compiler, so that if the definition of >DOES changes, it will still
work. (Not currently planning a different >DOES, but found this potential
problem while trying one.)

Add bracket-create.fs, so that (CREATE) can be POSTPONEd by CREATE, (about
to use this for SMite).

Reorder bracket-does.fs’s INCLUDE in make-base.fs so that (DOES) can also be
POSTPONEd during compilation (not used, but I did try it).