Simplify RESOLVERs, allowing only one type
The previous implementation assumed the target address was aligned. For
SMite we need unaligned targets, so we can’t use the low-order bit as a type
bit.

Rather than try to fix this, just remove it for now by having the target
define RESOLVER-BRANCH which is either BRANCH or CALL.

Hence, RESOLVER can no longer be used for general-purpose redefinition, but
that’s fine, as we already moved all of that code into
extras/redefinition.fs and it’s not currently used.