Generalize design of initialization code
Have a platform-provided initialization word or words to do
platform-specific initialization, which must end by calling START with the
value of LIMIT.

This design allows some of the code to be assembly (this is needed on RISC
OS), and the rest high level. The last word defined is the one used for the
initial branch.