RISC OS: fix a bug in 'THROW!
In a near-miracle, 'THROW!, which had no return from subroutine, was
executing its way harmlessly through the header of (U/MOD) on startup, and
then returning.

The bug turned up when I changed the code in INITIALIZE slightly, which set
condition flags that caused a non-existent SWI to be attempted (the first
cell of the name field of (U/MOD)).