Add command-line parsing to RISC OS
Because the actual parsing code needs pForth to be initialized, so it can
compile values into the dictionary, add another hook, PARSE-COMMAND-LINE to
START, and supply empty definitions for the other platforms.

Perhaps this should be generalized to post-setup initialization.

At the same time, make RISC OS’s init code slightly higher level and add a
couple of comments.

 The parser simply parses space-separated strings, and has no support for
 escapes or quoting.