src/compiler4.fs: add an info field to anonymous (:NONAME) words
COMPILE, on some platforms (Beetle and SMite) requires the existence of this
field to check whether the word should be inlined. Before, the relevant byte
happened always to be zero, at least for the very few cases in which
anonymous words were used, but a change to SMite’s primitives violated this
invariant for S".

There’s no reason why it shouldn’t be possible to check the info field for
any execution token, so add it before every :NONAME word.