Fix code-generation bug caused by running wrong NOPALIGN in DOES>
For now, not building DOES> in the meta-compiler fixes the bug. This may
fail to work if the definition of DOES> is changed in future, so leave the
line that compiles does.fs in the meta-compiler commented out as a FIXME.

The problem is that DOES> calls NOPALIGN, which now must be defined
differently on Beetle and SMite. When Beetle NOPALIGN runs on SMite, it
compiles extra zero bytes, which on SMite cause the constant 0 to be pushed
on to the stack when the definition of DEFER compiled with the meta-compiler
runs.