SMite: use NUL-terminated strings for file EXTRA instructions
In fact, the reading of counted strings was recently broken in SMite; as a
result, I changed it to require NUL-terminated strings.

Use C0END to supply them; not that this means that RENAME-FILE is now broken
on SMite as well as on RISC OS (which I hadn’t previously noticed). Fix
coming after this.