configure.ac: overhaul behaviour of setting host and build architectures
Require --host or --build to be set if corresponding BUILD_EXECUTOR or
HOST_EXECUTOR is (as it’s not in general possible to determine the arch from
the name of the executor).

Check that the build executor can run the corresponding pforth-base image.

If HOST_EXECUTOR is specified and does not exist as a file on the build
system, emit a warning.