highlevel.fs.in: fix (THROW) to restore data stack correctly (fix #18)
When SP is saved in CATCH, there is an xt on the stack that is not required
when the stack is restored, so add a DROP to (THROW).

I found the bug by comparing with the sample Forth 2012 implementation of
CATCH and THROW, and made pForth conform more closely, by renaming 'FRAME to
HANDLER, and by reordering the information saved in the exception frame on
the return stack; however, the sample implementation does not save the input
source.