highlevel.fs: fix out-by-one error in CATCH/(THROW)
The value of SP saved and restored was out by a cell. This did not matter on
targets where SP is an address, since (THROW) compensated with a DROP.
However on mit where SP is really the depth of the stack, the corresponding
SP! did not work, since it was required to increase the depth, which it
cannot. Hence, it was crashing.