Re-enable image saving, reintroducing SAVE-IMAGE
The resulting image is not relocatable in general, as any definitions added
to the original base image will not be relocatable.

Do not add the “extended image” back to the build system, as this was a lot
of extra complexity without much extra utility.

The ability to restart an pForth image that has had extra definitions added
is required by Mit’s test suite.