RISC OS: fix compilation of SAVE-OBJECT
The meta-compiler’s version of SAVE-OBJECT was being built without access to
TYPE-FILE. As a result, the file type of pforth-new was not set, and the
triple test did not run.

However, the error was not detected, because *Run of a data file seems to
cause an uncatchable error, so that !TestBuild was exiting at this point but
not returning an error. Add the IfType utility and use it to catch the file
type not being set correctly. Also improve the script so that it can detect
either the first-stage pforth-new or second-stage pforth-new not having its
file type correctly set.