Beetle: don’t use (LITERAL)I to compile address literals
This makes the build reproducible when it is built at different 'FORTH
addresses. For example, Mit tends to run in high memory on Linux, and
addresses require 4 bytes; whereas on Beetle, addresses fit in 3, and
a (LITERAL)I instruction can be used. This makes Beetle and Mit build
different binaries.