Fold LITERAL, into LITERAL; replace ADDRESS-LITERAL with RELATIVE-LITERAL
Do not require that a literal be compiled as a literal word.

Use the “DOES> offset” part of the header’s info field to navigate to the
DOES> code, thus allowing a relative offset to be easily compiled (forward
reference not needed).

Add and use RELATIVE-LITERAL and CURRENT-RELATIVE-LITERAL, replacing
ADDRESS-LITERAL.