highlevel.fs.in: fix a pair of stack bugs
There are two problems here: first, INTERPRET was trying to drop a
non-existent stack item, which caused a crash on Bee.

Commit d857ed4 introduced a fix for a related problem, but I failed to
notice then that the offset added to the saved value of SP must be
multiplied by STACK-DIRECTION, so on Beetle (which has a descending stack),
SP ended up 2 cells out rather than 1. This went unnoticed, because the
stack was immediately reset on error by (QUIT), and furthermore added enough
phantom stack items that the erroneous NIP in INTERPRET worked.