Fix build system to build both 32-bit and 64-bit pForth images
Check in pforth.s (as well as retaining the pForth images as
pforth-{32,64}.bin). Use bee-as to build the non-native image, and build the
native image with the usual triple-test method.

Also, use the new built-in constants bee_word_{bits,bytes} in bee-as,
instead of passing a --defsym flag to as.